Perf fixes for quantized DiT (Krea2/Anima) on Turing + keep-resident decode
- memory_management: force full load for KModel (int4/int8 DiT), fix >= free_memory boundary,
  add keep_all_existing param to load_models_gpu (skip unloading resident models)
- patcher/vae: decode memory estimate 64x->8x; keep resident DiT while loading VAE (mini-aimdo)
- operations: compute dtype fp32->fp16 on Turing (no bf16 support, fp32 halves throughput)
- patcher/base: LoRA weight backup forced to CPU (HIGH_VRAM), INT8 custom layers use
  weight_function online merge instead of patch_weight_to_device

+++ backend/operations.py
@@ -644,7 +644,7 @@ def using_forge_operations(operations=None, device=None, dtype=None, manual_cast
         # https://github.com/Comfy-Org/ComfyUI/blob/v0.16.4/comfy/ops.py#L950
 
         _device = memory_management.get_torch_device()
-        _dtype = torch.bfloat16 if memory_management.should_use_bf16(_device) else torch.float32
+        _dtype = torch.bfloat16 if memory_management.should_use_bf16(_device) else torch.float16  # [0902] fp32->fp16: Turing 无 bf16, fp32 算力减半; ComfyUI 同卡 fp16=1.22 it/s
         fp8_compute = memory_management.supports_fp8_compute(_device)
         nvfp4_compute = memory_management.supports_nvfp4_compute(_device)
         mxfp8_compute = memory_management.supports_mxfp8_compute(_device)
